% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smpic_dim}
\alias{smpic_dim}
\title{Overview of Image Dimensions for Social Media}
\format{A data frame with 32 rows and 6 variables:
\describe{
  \item{media}{the social media in question}
  \item{picture}{the type of image in question}
  \item{width}{the width of image in question}
  \item{height}{the height of image in question}
  \item{dimension}{the dimension of image in question - height / width}
  \item{id}{the id of image in question}
}}
\source{
\url{https://makeawebsitehub.com/social-media-image-sizes-cheat-sheet/}
}
\usage{
smpic_dim
}
\description{
This data set contains dimensions for 32 different images sizes for most
social media sites
}
\keyword{datasets}
